TS: extracting PaginationItem from Table/Footer
